Tidies up bootstrap and loggers.
